feat: Switch to double.

+++ scripts/chunk/ChunkColumn.cs
@@ -5,18 +5,18 @@ using System.Linq;
 public class ChunkColumn
 {
     public Vector2I Position;
-    public float[,] HeightMap;
-    public float HeightMapHigh;
-    public float HeightMapLow;
+    public double[,] HeightMap;
+    public double HeightMapHigh;
+    public double HeightMapLow;
     public readonly ConcurrentDictionary<Vector3I, Chunk> Chunks = new();
 
     private ChunkColumn() { }
-    public ChunkColumn(Vector2I position, float[,] heightMap)
+    public ChunkColumn(Vector2I position, double[,] heightMap)
     {
         Position = position;
         HeightMap = heightMap;
 
-        HeightMapHigh = heightMap.Cast<float>().Max();
-        HeightMapLow = heightMap.Cast<float>().Min();
+        HeightMapHigh = heightMap.Cast<double>().Max();
+        HeightMapLow = heightMap.Cast<double>().Min();
     }
 }
